+++ CHANGES.md
@@ -19,6 +19,9 @@ New features:
   - A new pool connection parameter `?max_idle_age` provides removal
     resources from pools which have been unused for the given period.
 
+  - Printf-style function `Caqti_query.qprintf` for dynamic queries
+    (GPR#103, Basile Clément).
+
 Breaking changes:
 
   - The minimal OCaml requirement is now 4.08.0.